+++ src/upsimulator/rules/conditions/PromoterCondition.java
@@ -25,7 +25,7 @@ public class PromoterCondition extends PObject implements Condition {
 
 	@Override
 	public int satisfy(Membrane membrane) {
-		if (membrane.getNumOf(this) > 0) {
+		if ((Integer)membrane.getNumOf(this) > 0) {
 			return Integer.MAX_VALUE;
 		} else
 			return 0;
